requirements.txt
telethon
python-dotenv
psycopg2-binary
sqlalchemy
dbt-postgres
fastapi
uvicorn
pydantic
ultralytics
dagster
dagster-webserver
